requirements.txt
pyyaml>=6.0
python-dotenv>=0.19.0
ffmpeg-python>=0.2.0
typing-extensions>=4.0.0
colorlog>=6.7.0
click>=8.0.0